Some docs clean up
