+++ README.md
@@ -17,7 +17,7 @@ Additional requirements for creating React Native projects:
 
 ## Usage
 
-The PX Blue CLI can be utilized via npx without having to install any global packages:
+The PX Blue CLI can be utilized via npx without having to install any global packages (recommended):
 
 ```
 npx -p @pxblue/cli pxb <command>
@@ -50,7 +50,7 @@ The following table list out some options for the `pxb new` command. All these o
 | Option                                                         | Description                                                                                                                                                                                                                           |
 | -------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
 | <code>--framework=<angular\|react\|ionic\|react-native></code> | The framework in which the project will be generated.                                                                                                                                                                                 |
-| ```--name=<name>```                                            | Project name                                                                                                                                                                                                                          |
+| `--name=<name>`                                                | Project name                                                                                                                                                                                                                          |
 | <code>--cli=<rnc\|expo></code>                                 | (React Native projects only) which CLI to use to generate the project. We support `rnc` ([React-Native Community CLI](https://github.com/react-native-community/cli)) or `expo` ([Expo CLI](https://docs.expo.io/workflow/expo-cli/)) |
 | `--lint`                                                       | (TypeScript projects only) Install and configure PX Blue lint and prettier packages (omit or `--lint=false` to disable)                                                                                                               |
 | <code>--language=<typescript\|javascript></code>               | (React & React Native Only) The language in which the project will be generated                                                                                                                                                       |
